\name{samplingbook-package}
\Rdversion{1.1}
\alias{samplingbook-package}
\alias{samplingbook}
\docType{package}
\title{
Survey Sampling Procedures
}
\description{
Sampling procedures from the book 'Stichproben - Methoden und praktische Umsetzung mit R' by Goeran Kauermann and Helmut Kuechenhoff (2010).
}
\details{
\tabular{ll}{
Package: \tab Samplingbook\cr
Type: \tab Package\cr
Version: \tab 1.2.1\cr
Date: \tab 2016-07-05\cr
License: \tab GPL(>=2)\cr
LazyLoad: \tab yes\cr
}
Index:
\tabular{ll}{
election\tab German Parliament Election Data\cr
htestimate\tab Horvitz-Thompson Estimator\cr
influenza\tab Population and Cases of Influenza in Administrative Districts of Germany\cr
mbes\tab Model Based Estimation\cr
money\tab Money Data Frame\cr
pop\tab Small Suppositious Sampling Example\cr
pps.sampling \tab Sampling with Probabilities Proportional to Size\cr
sample.size.mean \tab Sample Size Calculation for Mean Estimation\cr
sample.size.prop \tab Sample Size Calculation for Proportion Estimation\cr
Samplingbook-package \tab Survey Sampling Procedures\cr
Smean \tab Sampling Mean Estimation\cr
Sprop \tab Sampling Proportion Estimation\cr
stratamean \tab Stratified Sample Mean Estimation\cr
stratasamp \tab Sample Size Calculation for Stratified Sampling\cr
tax \tab Hypothetical Tax Refund Data Frame\cr
}
}
\author{
Author: Juliane Manitz <r@manitz.org>,
  contributions by \cr
  Mark Hempelmann <mark.hempelmann@o2online.de>, \cr
  Goeran Kauermann <gkauermann@wiwi.uni-bielefeld.de>, \cr
  Helmut Kuechenhoff <kuechenhoff@stat.uni-muenchen.de>, \cr
	Shuai Shao <shuai.shao@campus.lmu.de>, \cr
  Cornelia Oberhauser <conny.oberhauser@gmx.de>, \cr
  Nina Westerheide <nwesterheide@wiwi.uni-bielefeld.de>, \cr
  Manuel Wiesenfarth <m.wiesenfarth@uni-goettingen.de> \cr\cr
Maintainer: Juliane Manitz  <r@manitz.org>
}
\references{
  Kauermann, Goeran/Kuechenhoff, Helmut (2010): Stichproben. Methoden und praktische Umsetzung mit R. Springer.
}
